+++ src/pychart/runner.py
@@ -1,5 +1,5 @@
-from src.scanner import Scanner
-from src.pyparser import Parser
+from .scanner import Scanner
+from .pyparser import Parser
 
 
 def run(source: str):
