+++ src/pychart/token_type/token.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 from typing import Any
-from src.token_type.token_type_enum import TokenType
+from .token_type_enum import TokenType
 
 
 @dataclass(frozen=True)
